Sends JPEG in packets now
ReadSD function modified to send JPEG in 32 byte packets to Serial. 
ReadSD first sends "Reading" and the number of bytes to be read over the Serial. 
Processing code modified to print out all Serial notifications to the console, to recognize the "Reading" signal and the number of bytes to be read, and to return to the initial state of reading everything to the console after the file has been read.